# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FightFind API
  version: 2.0.0
  description: |
    FightFind is a combat-sports marketplace and network for India. This API serves
    the fighter mobile app, the gym owner portal and the Razorpay webhook pipeline.

    Conventions:
    - All money is integer **paise** (`₹999 = 99900`). The server always resolves prices.
    - Cursor pagination returns `{ items, nextCursor }`.
    - Errors use `{ "error": { "code", "message", "details?", "requestId" } }`.
servers:
  - url: http://localhost:4001
    description: Local development
  - url: https://api.fightfind.in
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/otp/request:
    post:
      tags: [Auth]
      security: []
      summary: Request a phone or email OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, identifier]
              properties:
                channel: { type: string, enum: [phone, email] }
                identifier: { type: string, example: "+919876543210" }
                purpose:
                  type: string
                  enum: [login, signup, gym_owner_login, gym_owner_signup]
                  default: login
                name: { type: string }
      responses:
        "200":
          description: OTP sent. `devOtp` is only returned outside production.
          content:
            application/json:
              schema:
                type: object
                properties:
                  challengeId: { type: string, format: uuid }
                  channel: { type: string }
                  identifier: { type: string }
                  expiresAt: { type: string, format: date-time }
                  devOtp: { type: string, description: Development only }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/otp/verify:
    post:
      tags: [Auth]
      security: []
      summary: Verify an OTP and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challengeId, code]
              properties:
                challengeId: { type: string, format: uuid }
                code: { type: string, pattern: "^\\d{6}$" }
                name: { type: string }
      responses:
        "200":
          description: Session created. The refresh token is also set as an httpOnly cookie.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "400": { $ref: "#/components/responses/ApiError" }
...
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate a refresh token (body or httpOnly cookie)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
        "401": { $ref: "#/components/responses/ApiError" }
...
  /api/v1/subscriptions/gym-platform:
    post:
      tags: [Subscriptions, Gym Owner]
      summary: Start the gym platform plan (requires listing fee paid)
      responses:
        "201": { description: Subscription }
        "409": { description: LISTING_PAYMENT_REQUIRED }
...
  /api/v1/subscriptions/{id}:
    get:
      tags: [Subscriptions]
      summary: Subscription detail
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: "`{ subscription }`" }
  /api/v1/subscriptions/{id}/cancel:
    post:
      tags: [Subscriptions]
      summary: Cancel a subscription
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cancelAtPeriodEnd: { type: boolean, default: true }
      responses:
        "200": { description: Updated subscription }
  /api/v1/subscriptions/{id}/verify:
    post:
      tags: [Subscriptions]
      summary: Verify a subscription Checkout callback
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Verified }
...
  /api/v1/admin/gyms/{id}/verification:
    post:
      tags: [Admin]
      summary: Verify / reject / suspend a gym
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [pending, verified, rejected, suspended] }
                reason: { type: string }
      responses:
        "200": { description: "`{ gym }`" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ApiError:
      description: Structured error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
                  details: {}
                  requestId: { type: string }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, example: RATE_LIMITED }
                  message: { type: string }
  schemas:
    AuthSession:
      type: object
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
        tokens:
          type: object
          properties:
            accessToken: { type: string }
            accessTokenExpiresAt: { type: string, format: date-time }
            refreshToken: { type: string }
            refreshTokenExpiresAt: { type: string, format: date-time }
    AuthUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        email: { type: [string, "null"] }
        roles:
          type: array
          items: { type: string, enum: [FIGHTER, GYM_OWNER, ADMIN] }
        isGuest: { type: boolean }
        onboarding:
          type: object
          properties:
            hasFighterProfile: { type: boolean }
            fighterProfileComplete: { type: boolean }
            isGymOwner: { type: boolean }
            gymCount: { type: integer }
...